Created a separate search filter component for staff dashboard
